Module icu_codepointtrie::codepointtrie

Structs

CodePointTrie

This struct represents a de-serialized CodePointTrie that was exported from ICU binary data.

CodePointTrieHeader

This struct contains the fixed-length header fields of a CodePointTrie.

Enums

TrieType

The type of trie represents whether the trie has an optimization that would make it small or fast. See UCPTrieType in ICU4C.

Traits

TrieValue

A trait representing the values stored in the data array of a CodePointTrie. This trait is used as a type parameter in constructing a CodePointTrie.