Trait icu_codepointtrie::TrieValue

pub trait TrieValue: Copy + Eq + PartialEq + AsULE + 'static {
    type TryFromU32Error: Display;

    const DATA_GET_ERROR_VALUE: Self;

    fn try_from_u32(i: u32) -> Result<Self, Self::TryFromU32Error>;
}

A trait representing the values stored in the data array of a CodePointTrie. This trait is used as a type parameter in constructing a CodePointTrie.

Associated Types

type TryFromU32Error: Display

Error type when converting from a u32 to this TrieValue.

Associated Constants

const DATA_GET_ERROR_VALUE: Self

Last-resort fallback value to return if we cannot read data from the trie.

In most cases, the error value is read from the last element of the data array.

Required methods

fn try_from_u32(i: u32) -> Result<Self, Self::TryFromU32Error>

A parsing function that is primarily motivated by deserialization contexts. When the serialization type width is smaller than 32 bits, then it is expected that the call site will widen the value to a u32 first.